cleaning up the types a bit
